Address code review: extract, unify, simplify scheduling API
Structural:
- Extract entity scheduling (Do, DoAfter, schedule, scheduleAfter,
  runScheduled, signal helpers) from entity.go into entity_schedule.go.
  entity.go drops from 653 to 516 lines.
- Unify task completion: add executeWithRecovery helper used by both
  scheduledTransaction.Run and EntityHandle.runScheduled, eliminating
  duplicated recover/PanicError logic.
- Collapse Call/CallEntity/CallRef wait logic into awaitTask[T] helper.
- Extract assertEntity[T] helper for EntityRef type-assertion
  boilerplate.

API cleanup:
- Remove Context.World() (returned *Tx, name lied). All callers use
  Tx() directly.
- Standardize Player.Do/DoAfter callback signature to (ctx, p),
  matching EntityRef's (ctx, entity) convention.
- Player.Do/DoAfter nil path now returns NewFinishedTask(ErrEntityClosed)
  instead of constructing a no-op callback.
- Death path uses player.NewRef typed ref instead of raw p.H().DoAfter
  with an unsafe type assert.
- Use errors.AsType (Go 1.26) instead of errors.As in server.go.

Behavioral:
- Respawn scheduling failure (destination world closed) now closes the
  orphaned entity handle and disconnects the session instead of silently
  losing the player.
