+++ server/world/context.go
@@ -26,11 +26,6 @@ func newContext(tx *Tx) *Context {
 // only valid for the duration of the callback.
 func (ctx *Context) Tx() *Tx { return ctx.Val() }
 
-// World returns the owner-scoped world access for the context. It returns the
-// current transaction rather than the raw *World, so callers can use world
-// operations without re-entering World.Do.
-func (ctx *Context) World() *Tx { return ctx.Tx() }
-
 // Defer schedules f to run after the current owner callback completes.
 func (ctx *Context) Defer(f func(ctx *Context)) *Task {
 	return ctx.Tx().deferTask(func(ctx *Context) error {
